Simplify: extract useChannelMetaName from VideoPlayerOverlayImpl (#16 slice 1)
First decomposition slice of the 3,224-line VideoPlayerOverlayImpl god-component.
Move the self-contained channel-meta lookup (module-level TTL cache + the one
effect that resolves a channel's display name) into lib/useChannelMetaName.ts as
a custom hook; the component now calls `useChannelMetaName(currentVideo, rpc)`.

Behavior preserved: code moved verbatim, same module-level cache semantics, same
effect dep array (keyed on channelKey only, with the exhaustive-deps disable made
explicit). Replacing a useState+useEffect pair with one unconditional hook call
keeps hook order valid; channelMetaName is read only in render, so effect-order
is unaffected. Component 3,224 -> 3,183 lines.

Validated locally: app `tsc --noEmit` reports zero new type errors in the changed
files (the one pre-existing absoluteFillObject error is unrelated); hook structure
satisfies rules-of-hooks (CI lint gate covers react-hooks).
